+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## ????
 - Default headers: connection, accept, user-agent
 - Can provide options on each verb method
+- Support HTTP requests without curl, using file_get_contents()
 
 ## 0.0.1 : 2023-12-02
 - Basic working version